git: ensure architecture independent files are installed appropriately.
Git's build system is an utter piece of crap. We "make" the configure
script, then run it so that we can pass --exec-prefix and --prefix
appropriately. Then we can proceed with makeing as usual, but the perl
and python path argument forms differ in each step, so handle that. And
git *still* doesn't handle the perl pod files appropriately, so
fixlocalpod() from perl-module doesn't work anymore.

Change-Id: Id24eb579d13b6e335b4db2f360e61232cc17b447